Move DataEntryAdapter.kt and auxiliar classes to module
