+++ app/src/androidTest/java/org/dhis2/usescases/datasets/DataSetInitialRobot.kt
@@ -3,17 +3,13 @@ package org.dhis2.usescases.datasets
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.contrib.RecyclerViewActions
-import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.hasDescendant
 import androidx.test.espresso.matcher.ViewMatchers.hasSibling
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import org.dhis2.R
 import org.dhis2.common.BaseRobot
-import org.dhis2.common.viewactions.typeChildViewWithId
-import org.dhis2.data.forms.dataentry.fields.FormViewHolder
 import org.dhis2.utils.customviews.DateViewHolder
-import org.hamcrest.Matchers
 import org.hamcrest.Matchers.allOf
