+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2021-12-10 11:46","gitSha":"e6b6755665"}
+{"buildTime":"2022-01-12 16:38","gitSha":"f209e8bfe"}
