--- app/src/main/java/org/dhis2/data/forms/dataentry/fields/EditableFieldViewModel.java
@@ -1,5 +1,0 @@
-package org.dhis2.data.forms.dataentry.fields;
-
-public abstract class EditableFieldViewModel<T> extends FieldViewModel {
-
-}
