+++ app/src/test/java/org/dhis2/data/forms/dataentry/SectionHandlerTest.kt
@@ -1,5 +1,6 @@
 package org.dhis2.data.forms.dataentry
 
+import org.dhis2.form.ui.SectionHandler
 import org.junit.Assert.assertTrue
 import org.junit.Test
