+++ form/src/main/java/org/dhis2/form/ui/SectionHandler.kt
@@ -1,4 +1,4 @@
-package org.dhis2.data.forms.dataentry
+package org.dhis2.form.ui
 
 class SectionHandler {
     fun getSectionPositionFromVisiblePosition(
