+++ form/src/main/java/org/dhis2/form/ui/style/FormUiModelColorFactoryImpl.kt
@@ -1,10 +1,9 @@
-package org.dhis2.data.forms.dataentry
+package org.dhis2.form.ui.style
 
 import android.content.Context
 import androidx.core.content.ContextCompat
 import org.dhis2.commons.resources.ColorUtils
-import org.dhis2.form.ui.style.FormUiColorFactory
-import org.dhis2.form.ui.style.FormUiColorType
+import org.dhis2.form.R
 
 class FormUiModelColorFactoryImpl(
     val context: Context,
@@ -16,15 +15,15 @@ class FormUiModelColorFactoryImpl(
                 FormUiColorType.PRIMARY to
                     ColorUtils.getPrimaryColor(context, ColorUtils.ColorType.PRIMARY),
                 FormUiColorType.TEXT_PRIMARY to
-                    ContextCompat.getColor(context, org.dhis2.R.color.textPrimary),
+                    ContextCompat.getColor(context, R.color.textPrimary),
                 FormUiColorType.FIELD_LABEL_TEXT to
-                    ContextCompat.getColor(context, org.dhis2.R.color.text_black_A63),
+                    ContextCompat.getColor(context, R.color.text_black_A63),
                 FormUiColorType.WARNING to
-                    ContextCompat.getColor(context, org.dhis2.R.color.warning_color),
+                    ContextCompat.getColor(context, R.color.warning_color),
                 FormUiColorType.ERROR to
-                    ContextCompat.getColor(context, org.dhis2.R.color.error_color),
+                    ContextCompat.getColor(context, R.color.error_color),
                 FormUiColorType.ACTION_ICON to
-                    ContextCompat.getColor(context, org.dhis2.R.color.colorGreyDefault)
+                    ContextCompat.getColor(context, R.color.colorGreyDefault)
             )
         }
         return mapOf(
@@ -35,9 +34,9 @@ class FormUiModelColorFactoryImpl(
             FormUiColorType.FIELD_LABEL_TEXT to
                 ColorUtils.getPrimaryColor(context, ColorUtils.ColorType.ACCENT),
             FormUiColorType.WARNING to
-                ContextCompat.getColor(context, org.dhis2.R.color.warning_color),
+                ContextCompat.getColor(context, R.color.warning_color),
             FormUiColorType.ERROR to
-                ContextCompat.getColor(context, org.dhis2.R.color.error_color),
+                ContextCompat.getColor(context, R.color.error_color),
             FormUiColorType.ACTION_ICON to
                 ColorUtils.getPrimaryColor(context, ColorUtils.ColorType.ACCENT)
         )
