+++ form/src/main/res/values/styles.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+
     <style name="FieldLabel">
         <item name="android:textSize">@dimen/field_label_size</item>
         <item name="android:layout_marginStart">16dp</item>
@@ -18,9 +19,14 @@
         <item name="android:textColor">@color/textPrimary</item>
         <item name="colorControlActivated">?colorPrimary</item>
     </style>
+
     <style name="dataInputTextAccent" parent="TextAppearance.AppCompat">
         <item name="android:textColorHint">@color/colorAccent</item>
         <item name="colorAccent">@color/colorAccent</item>
         <item name="colorControlActivated">@color/colorAccent</item>
+    </style>
+
+    <style name="searchFormInputText" parent="Theme.MaterialComponents.Light.NoActionBar.Bridge">
+        <item name="colorControlActivated">?colorAccent</item>
     </style>
 </resources>
